requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-cors-headers==4.2.0
django-import-export==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
install==1.3.5
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
psycopg2==2.9.6
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
